pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autonomous-file-intelligence"
version = "0.1.0"
description = "Resumable, safe, high-volume file intelligence and organization engine."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openai>=1.76.0",
  "tqdm>=4.66.0",
  "pypdf>=5.4.0",
  "python-docx>=1.1.0",
  "Pillow>=10.4.0",
  "pytesseract>=0.3.10",
]

[project.scripts]
file-intel = "file_intelligence.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
